fixup! feat(core): add mock RNGs for Optiga and Tropic
